attendence/requirements.txt
opencv-python
dlib
face-recognition
Flask
numpy
Pillow
cmake==3.28.1
keyboard
